backend/requirements.txt
flask
flask-cors
sentence-transformers
faiss-cpu
spacy
nltk
pypdf
numpy
python-dotenv
tqdm
groq
pyspellchecker
pytesseract
pdf2image
pillow
python-dotenv
